update sampling style
Samples are now individual rays, with images acting as a visualisation of the overall distribution.
Works best when no rays are duplicated, otherwise you have to adjust the intensity of each new ray by the probability of sampling it.
Sometimes the gradient messes up, in which case the joint score is currently replaced with a 0.